Split SapSystemsOverviePage and SaptuneDetailsPage
