+++ assets/js/pages/SaptuneDetailsPage/SaptuneDetails.test.jsx
