+++ assets/js/pages/SaptuneDetailsPage/SaptuneDetailsPage.jsx
