+++ assets/js/pages/SaptuneDetailsPage/SaptuneDetailsPage.test.jsx
