+++ assets/js/pages/SaptuneDetailsPage/SaptuneVersion.jsx
