+++ assets/js/pages/SaptuneDetailsPage/SaptuneVersion.test.jsx
