+++ assets/js/pages/SaptuneDetailsPage/index.js
